#8: Refacting cli tool for multi stage jobs
